Html: Move block/html.js, to deserializeHtml.js
