--- src/html/blocks/html.js
@@ -1,17 +1,0 @@
-const { Deserializer } = require('../../');
-const parse = require('./parse');
-
-/**
- * Deserialize an HTML string
- * @type {Deserializer}
- */
-const deserialize = Deserializer()
-.then((state) => {
-    const nodes = parse(state.text);
-    return state
-        .push(nodes)
-        .skip(state.text.length);
-});
-
-module.exports = { deserialize };
-
