+++ src/html/blocks/index.js
@@ -12,6 +12,5 @@ module.exports = [
     require('./table'),
     require('./row'),
     require('./cell'),
-    require('./footnote'),
-    require('./html')
+    require('./footnote')
 ];
